TODO: enable MDEV-13049 optimization for 10.3

+++ sql/opt_sum.cc
@@ -768,12 +768,12 @@ static bool matching_cond(bool max_fl, TABLE_REF *ref, KEY *keyinfo,
   key_part_map org_key_part_used= *key_part_used;
   if (eq_type || between || max_fl == less_fl)
   {
-    uint length= (key_ptr-ref->key_buff)+part->store_length;
+    uint length= (uint)(key_ptr-ref->key_buff)+part->store_length;
     if (ref->key_length < length)
     {
     /* Ultimately ref->key_length will contain the length of the search key */
       ref->key_length= length;      
-      ref->key_parts= (part - keyinfo->key_part) + 1;
+      ref->key_parts= (uint)(part - keyinfo->key_part) + 1;
     }
     if (!*prefix_len && part+1 == field_part)       
       *prefix_len= length;
